Test suite improvements: use tmpdir fixture and publicly expose environment test helper methods (#99)
* Use tmpdir fixture

* Allow test_envs suite to be run against any set of environments

* Make tests part of package

* Revert "Make tests part of package"

This reverts commit c4920b705eb172089aa04fd4d3a8141aa0980208.

* Split helper methods into separate testing module as well

* Remove dead code
