+++ src/imitation/testing/__init__.py
@@ -1,0 +1,1 @@
+"""Helper methods for unit tests. May also be useful for users of imitation."""
