+++ tests/test_policies.py
@@ -1,7 +1,5 @@
 """Test imitation.policies."""
 
-import tempfile
-
 import numpy as np
 import pytest
 from stable_baselines.common.vec_env import VecNormalize
@@ -34,7 +32,7 @@ def test_actions_valid(env_name, policy_type):
 @pytest.mark.parametrize("env_name", SIMPLE_ENVS)
 @pytest.mark.parametrize("model_cfg", BASELINE_MODELS)
 @pytest.mark.parametrize("normalize", [False, True])
-def test_serialize_identity(env_name, model_cfg, normalize):
+def test_serialize_identity(env_name, model_cfg, normalize, tmpdir):
   """Test output actions of deserialized policy are same as original."""
   orig_venv = venv = util.make_vec_env(env_name, n_envs=1, parallel=False)
   vec_normalize = None
@@ -59,15 +57,14 @@ def test_serialize_identity(env_name, model_cfg, normalize):
   orig_rollout = rollout.generate_transitions(model, venv, n_timesteps=1000,
                                               deterministic_policy=True)
 
-  with tempfile.TemporaryDirectory(prefix='imitation-serialize-pol') as tmpdir:
-    serialize.save_stable_model(tmpdir, model, vec_normalize)
-    # We use `orig_venv` since `load_policy` automatically wraps `loaded`
-    # with a VecNormalize, when appropriate.
-    with serialize.load_policy(model_name, tmpdir, orig_venv) as loaded:
-      orig_venv.env_method('seed', 0)
-      orig_venv.reset()
-      new_rollout = rollout.generate_transitions(loaded, orig_venv,
-                                                 n_timesteps=1000,
-                                                 deterministic_policy=True)
+  serialize.save_stable_model(tmpdir, model, vec_normalize)
+  # We use `orig_venv` since `load_policy` automatically wraps `loaded`
+  # with a VecNormalize, when appropriate.
+  with serialize.load_policy(model_name, tmpdir, orig_venv) as loaded:
+    orig_venv.env_method('seed', 0)
+    orig_venv.reset()
+    new_rollout = rollout.generate_transitions(loaded, orig_venv,
+                                               n_timesteps=1000,
+                                               deterministic_policy=True)
 
   assert np.allclose(orig_rollout.act, new_rollout.act)
