+++ tests/test_scripts.py
@@ -6,7 +6,6 @@ named_config for each experiment implicitly sets parallel=False.
 """
 
 import os.path as osp
-from tempfile import TemporaryDirectory
 
 import pytest
 import ray.tune as tune
@@ -17,30 +16,29 @@ from imitation.scripts.parallel import parallel_ex
 from imitation.scripts.train_adversarial import train_ex
 
 
-def test_expert_demos_main():
+def test_expert_demos_main(tmpdir):
   """Smoke test for imitation.scripts.expert_demos.rollouts_and_policy"""
-  with TemporaryDirectory(prefix='imitation-data_collect-main') as tmpdir:
-    run = expert_demos_ex.run(
-        named_configs=['cartpole', 'fast'],
-        config_updates=dict(
-          log_root=tmpdir,
-        ),
-    )
-    assert run.status == 'COMPLETED'
-    assert isinstance(run.result, dict)
+  run = expert_demos_ex.run(
+      named_configs=['cartpole', 'fast'],
+      config_updates=dict(
+        log_root=tmpdir,
+      ),
+  )
+  assert run.status == 'COMPLETED'
+  assert isinstance(run.result, dict)
 
 
-def test_expert_demos_rollouts_from_policy():
+def test_expert_demos_rollouts_from_policy(tmpdir):
   """Smoke test for imitation.scripts.expert_demos.rollouts_from_policy"""
-  with TemporaryDirectory(prefix='imitation-data_collect-policy') as tmpdir:
-    run = expert_demos_ex.run(
-        command_name="rollouts_from_policy",
-        named_configs=['cartpole', 'fast'],
-        config_updates=dict(
-          log_root=tmpdir,
-          rollout_save_path=osp.join(tmpdir, "rollouts", "test.pkl"),
-          policy_path="tests/data/cartpole_0/policies/final/",
-        ))
+  run = expert_demos_ex.run(
+      command_name="rollouts_from_policy",
+      named_configs=['cartpole', 'fast'],
+      config_updates=dict(
+        log_root=tmpdir,
+        rollout_save_path=osp.join(tmpdir, "rollouts", "test.pkl"),
+        policy_path="tests/data/cartpole_0/policies/final/",
+      ),
+  )
   assert run.status == 'COMPLETED'
 
 
@@ -51,71 +49,67 @@ EVAL_POLICY_CONFIGS = [
 
 
 @pytest.mark.parametrize('config', EVAL_POLICY_CONFIGS)
-def test_eval_policy(config):
+def test_eval_policy(config, tmpdir):
   """Smoke test for imitation.scripts.eval_policy"""
-  with TemporaryDirectory(prefix='imitation-policy_eval') as tmpdir:
-      config_updates = {
-          'render': False,
-          'log_root': tmpdir,
-      }
-      config_updates.update(config)
-      run = eval_policy_ex.run(config_updates=config_updates,
-                               named_configs=['fast'])
-      assert run.status == 'COMPLETED'
-      assert isinstance(run.result, dict)
+  config_updates = {
+      'render': False,
+      'log_root': tmpdir,
+  }
+  config_updates.update(config)
+  run = eval_policy_ex.run(config_updates=config_updates,
+                           named_configs=['fast'])
+  assert run.status == 'COMPLETED'
+  assert isinstance(run.result, dict)
 
 
-def test_train_adversarial():
+def test_train_adversarial(tmpdir):
   """Smoke test for imitation.scripts.train_adversarial"""
-  with TemporaryDirectory(prefix='imitation-train') as tmpdir:
-    named_configs = ['cartpole', 'gail', 'fast', 'plots']
-    config_updates = {
-        'init_trainer_kwargs': {
-            # Rollouts are small, decrease size of buffer to avoid warning
-            'trainer_kwargs': {
-                'n_disc_samples_per_buffer': 50,
-            },
-        },
-        'log_root': tmpdir,
-        'rollout_glob': "tests/data/cartpole_0/rollouts/final.pkl",
-    }
-    run = train_ex.run(
-        named_configs=named_configs,
-        config_updates=config_updates,
-    )
-    assert run.status == 'COMPLETED'
-    assert isinstance(run.result, dict)
+  named_configs = ['cartpole', 'gail', 'fast', 'plots']
+  config_updates = {
+      'init_trainer_kwargs': {
+          # Rollouts are small, decrease size of buffer to avoid warning
+          'trainer_kwargs': {
+              'n_disc_samples_per_buffer': 50,
+          },
+      },
+      'log_root': tmpdir,
+      'rollout_glob': "tests/data/cartpole_0/rollouts/final.pkl",
+  }
+  run = train_ex.run(
+      named_configs=named_configs,
+      config_updates=config_updates,
+  )
+  assert run.status == 'COMPLETED'
+  assert isinstance(run.result, dict)
 
 
-def test_transfer_learning():
+def test_transfer_learning(tmpdir):
   """Transfer learning smoke test.
 
   Save a dummy AIRL test reward, then load it for transfer learning."""
-
-  with TemporaryDirectory(prefix='imitation-transfer') as tmpdir:
-    log_dir_train = osp.join(tmpdir, "train")
-    run = train_ex.run(
-        named_configs=['cartpole', 'airl', 'fast'],
-        config_updates=dict(
+  log_dir_train = osp.join(tmpdir, "train")
+  run = train_ex.run(
+      named_configs=['cartpole', 'airl', 'fast'],
+      config_updates=dict(
           rollout_glob="tests/data/cartpole_0/rollouts/final.pkl",
           log_dir=log_dir_train,
-        ),
-    )
-    assert run.status == 'COMPLETED'
-    assert isinstance(run.result, dict)
+      ),
+  )
+  assert run.status == 'COMPLETED'
+  assert isinstance(run.result, dict)
 
-    log_dir_data = osp.join(tmpdir, "expert_demos")
-    discrim_path = osp.join(log_dir_train, "checkpoints", "final", "discrim")
-    run = expert_demos_ex.run(
-        named_configs=['cartpole', 'fast'],
-        config_updates=dict(
+  log_dir_data = osp.join(tmpdir, "expert_demos")
+  discrim_path = osp.join(log_dir_train, "checkpoints", "final", "discrim")
+  run = expert_demos_ex.run(
+      named_configs=['cartpole', 'fast'],
+      config_updates=dict(
           log_dir=log_dir_data,
           reward_type='DiscrimNet',
           reward_path=discrim_path,
-        ),
-    )
-    assert run.status == 'COMPLETED'
-    assert isinstance(run.result, dict)
+      ),
+  )
+  assert run.status == 'COMPLETED'
+  assert isinstance(run.result, dict)
 
 
 PARALLEL_CONFIG_UPDATES = [
@@ -128,7 +122,8 @@ PARALLEL_CONFIG_UPDATES = [
         "make_blank_policy_kwargs": {
           "learning_rate": tune.grid_search([3e-4, 1e-4]),
         },
-      }},
+      },
+    },
   ),
   dict(
     inner_experiment_name="train_adversarial",
@@ -142,7 +137,8 @@ PARALLEL_CONFIG_UPDATES = [
         },
         # Need absolute path because raylet runs in different working directory.
         'rollout_glob': osp.abspath("tests/data/cartpole_0/rollouts/final.pkl"),
-      }},
+      },
+    },
   ),
 ]
